device/requirements.txt
pyserial>=3.5
PyYAML>=6.0.1
requests>=2.31.0
